backend/requirements.txt
fastapi
uvicorn[standard]
python-socketio
pytest
pytest-asyncio
httpx
asgi-lifespan 
coverage
python-socketio[asyncio_client]
pydantic
orjson